Correct package names
This is how commit titles should be formed. The first line is the title. Title must be short, and be written in imperative form (Do something).
Then an empty line must follow, as above.
Then, one may provide additional description about the commit.
Titles start with capital letters and do not have dot at the end.

The imperative form is very nice when you natigate your project and want to see what this commit did. It is also
semantically very nice when you revert the commit.